requirements.txt
numpy==1.25.0
pandas==2.0.2
psycopg2_binary==2.9.6
PyPDF2==3.0.1
python-dotenv==1.0.0
scikit_learn==1.2.2
streamlit==1.24.0
transformers==4.30.2
black==23.3.0